requirements.txt
requirement.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.1
h11==0.14.0
idna==3.8
Jinja2==3.1.4
ldap3==2.9.1
MarkupSafe==2.1.5
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
fastapi
uvicorn
jinja2
python-multipart
ldap3
pyjwt
requests
openai
mysql-connector-python